pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "argus"
version = "1.0.0"
description = "Production ML observability platform with drift detection, root-cause analysis, and automated retraining"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
keywords = ["mlops", "drift-detection", "machine-learning", "monitoring", "fastapi"]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Science/Research",
    "Programming Language :: Python :: 3",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    # API
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.29.0",
    "python-multipart>=0.0.9",
    "httpx>=0.27.0",

    # ML
    "scikit-learn>=1.4.0",
    "numpy>=1.26.0",
    "pandas>=2.2.0",
    "scipy>=1.13.0",
    "joblib>=1.4.0",

    # Experiment tracking
    "mlflow>=2.12.0",

    # Data storage
    "pyarrow>=15.0.0",

    # Config
    "pyyaml>=6.0",
    "pydantic>=2.7.0",

    # Dashboard
    "streamlit>=1.33.0",
    "plotly>=5.20.0",
    "requests>=2.31.0",

    # Visualization (asset generation)
    "matplotlib>=3.8.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.4.0",
    "mypy>=1.9.0",
]

[project.scripts]
train = "scripts.train_initial_model:main"
simulate = "scripts.simulate_drift:main"
demo = "scripts.demo:main"

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = "-v --tb=short"
markers = [
    "selenium: end-to-end UI tests that require a running Streamlit dashboard",
]

[tool.ruff]
line-length = 100
target-version = "py310"
select = ["E", "F", "I", "W", "B", "C4", "UP"]
ignore = ["E501", "B008"]

[tool.mypy]
python_version = "3.10"
ignore_missing_imports = true
strict = false

[tool.hatch.build.targets.wheel]
packages = ["src"]
